Lazy-load IntervalSelect and Modal
